Remove gpg encryption of exported files
Remove the encryption step. It was necessary when we were planning
to use the prototype to export files to TNA's FTP server. Now that
we're just moving the files to another S3 bucket, we don't need to
encrypt the files with the TNA public key. We can do that later in
a separate step if we do need to copy the exported files to the
FTP site.