Download files to their original relative paths
When downloading the files, first look up their original paths in the TDR API,
and save the file from S3 to the original path rather than in the flat file
structure and UUID naming convention that we use in S3.

This means that when the files are exported, they are exported with their
original names.