Export consignment as BagIt bag
BagIt is a specification created by the US Library of Congress, and is
used to store and describe arbitrary collections of digital content.

It is a common archival format, so exporting a consignment as a Bag is
a flexible way for us to export the data, as opposed to saving the data
in the specific way required by our preservation system. We will have
to convert the BagIt data to a preservation SIP (Submission Information
Package) but we expect that to be straightforward.

Save a consignment as a BagIt bag which contains the consignment's files
and adds metadata about the consignment and the individual files, e.g.
the original client-side checksum.

For more information about BagIt, see:

- https://tools.ietf.org/html/rfc8493
- https://blogs.loc.gov/thesignal/2019/04/bagit-at-the-library-of-congress/